pyproject.toml
[tool.poetry]
name = "strava-segment-analyzer"
version = "0.1.0"
description = "A web app to analyze Strava segment efforts with heart rate and date filtering"
authors = ["Tom"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
flask = "2.3.3"
requests = "2.31.0"
python-dotenv = "1.0.0"
stravalib = "0.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
